Refactored SettingsWidget to remove connection management logic and replaced it with DatabasePage component.
