+++ mito-ai/src/Extensions/SettingsManager/database/types.ts
@@ -1,0 +1,20 @@
+export interface DBConnection {
+    type: string;
+    username: string;
+    password: string;
+    account: string;
+    warehouse: string;
+}
+
+export interface DBConnections {
+    [key: string]: DBConnection;
+}
+
+export interface NewConnectionForm {
+    name: string;
+    type: string;
+    username: string;
+    password: string;
+    account: string;
+    warehouse: string;
+}
